requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
faker
plotly